+++ example/lib/constants.dart
@@ -1,0 +1,17 @@
+const List<String> imagesList = [
+  'assets/images/colosseum.jpg',
+  'assets/images/florence.png',
+  'assets/images/Naples.png',
+  'https://lp-cms-production.imgix.net/2020-10/500pxRF_16641625.jpg',
+  'https://imageio.forbes.com/specials-images/imageserve/675172642/pura-ulun-danu-bratan-temple-in-Bali-/960x0.jpg?format=jpg&width=960',
+  'https://imageio.forbes.com/specials-images/imageserve/466090186/Panoramic-view-of-the-Mosques-of-Sultan-Hassan-and-Al-Rifai/960x0.jpg?format=jpg&width=960'
+];
+
+const List<String> titlesList = [
+  'Colosseum',
+  'Florence',
+  'Naples',
+  'Machu Picchu',
+  'Bali',
+  'Cairo'
+];