+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- Added parallax cards widget with parallax scrolling effect for horizontal and vertical scroll
+  views
+- Added option for custom card overlays and card properties
+- Added example app showcasing the use of the parallax cards widget